requirements.txt
tweepy>=4.14.0
pandas>=1.3.0
nltk>=3.6.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
flask>=2.0.0
pyyaml>=6.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
tqdm>=4.62.0

# Reddit API
praw>=7.6.0

# News API
requests>=2.27.1